+++ src/containers/Chat/ChatConversations/ChatConversations.tsx
@@ -5,8 +5,7 @@ import styles from './ChatConversations.module.css';
 import Loading from '../../../components/UI/Layout/Loading/Loading';
 import { SearchBar } from './SearchBar';
 import { GET_CONVERSATION_QUERY, FILTER_CONVERSATIONS_QUERY } from '../../../graphql/queries/Chat';
-import { MARK_AS_READ, MESSAGE_FRAGMENT } from '../../../graphql/mutations/Chat';
-import { useApolloClient, useLazyQuery, useMutation, gql } from '@apollo/client';
+import { useApolloClient, useLazyQuery, useMutation } from '@apollo/client';
 import selectedChatIcon from '../../../assets/images/icons/Chat/Selected.svg';
 
 export interface ChatConversationsProps {}
@@ -17,23 +16,6 @@ export const ChatConversations: React.SFC<ChatConversationsProps> = () => {
   const client = useApolloClient();
   const [searchVal, setSearchVal] = useState('');
   const [selectedIndex, setSelectedIndex] = useState(0); // This won't work all the time, if the default chat opened is not at index 0. (see ChatMessage for similar behavior).
-  const [markAsRead] = useMutation(MARK_AS_READ, {
-    onCompleted: (mydata) => {
-      mydata.markContactMessagesAsRead.map((messageId: any) => {
-        const message = client.readFragment({
-          id: `Message:${messageId}`,
-          fragment: MESSAGE_FRAGMENT,
-        });
-        const messageCopy = JSON.parse(JSON.stringify(message));
-        messageCopy.tags = messageCopy.tags.filter((tag: any) => tag.label !== 'Unread');
-        client.writeFragment({
-          id: `Message:${messageId}`,
-          fragment: MESSAGE_FRAGMENT,
-          data: messageCopy,
-        });
-      });
-    },
-  });
 
   const queryVariables = {
     contactOpts: {
@@ -84,13 +66,6 @@ export const ChatConversations: React.SFC<ChatConversationsProps> = () => {
     conversations = searchData.search.filter((n: any) => n.__typename === 'Conversation'); // Trying to only get conversation types from search query.
   }
 
-  const handleContactConversation = (index: number, contactId: number) => {
-    markAsRead({
-      variables: { contactId: contactId.toString() },
-    });
-    setSelectedIndex(index);
-  };
-
   // build the conversation list only if there are conversations
   let conversationList;
   if (conversations.length > 0) {
@@ -103,7 +78,7 @@ export const ChatConversations: React.SFC<ChatConversationsProps> = () => {
         <ChatConversation
           key={conversation.contact.id}
           selected={selectedIndex === index}
-          onClick={handleContactConversation}
+          onClick={(index: number) => setSelectedIndex(index)}
           index={index}
           contactId={conversation.contact.id}
           contactName={
